+++ src/js/const/asset.const.js
@@ -1,0 +1,4 @@
+export const ASSET_STATES = Object.freeze({
+  active: 0,
+  deleted: 1,
+})