+++ src/vue/modules/assets/asset-explorer/index.vue
@@ -39,7 +39,8 @@
       </template>
 
       <asset-attributes-viewer
-        :asset="selectedBalance.asset"
+        v-if="selectedBalance.asset"
+        :asset-code="selectedBalance.asset.code"
       />
 
       <div class="assets-explorer__actions">
@@ -62,7 +63,7 @@
         v-if="selectedBalance.asset"
         :asset-code="selectedBalance.asset.code"
         @submitted="(isAssetUpdateDrawerShown = false) ||
-          loadAccountBalances()"
+          loadAccountBalances() && loadAssets()"
       />
     </drawer>