+++ src/vue/modules/assets/shared/components/asset-attributes-viewer.vue
@@ -52,6 +52,12 @@
               </template>
             </td>
           </tr>
+          <tr v-if="asset.expirationDate">
+            <td>{{ 'assets.expiration-date-title' | globalize }}</td>
+            <td :title="asset.expirationDate | formatCalendar">
+              {{ asset.expirationDate | formatCalendar }}
+            </td>
+          </tr>
         </tbody>
       </table>
     </div>
@@ -68,7 +74,6 @@
 <script>
 import LogoViewer from './logo-viewer'
 
-import { AssetRecord } from '@/js/records/entities/asset.record'
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
@@ -78,7 +83,7 @@ export default {
     LogoViewer,
   },
   props: {
-    asset: { type: AssetRecord, required: true },
+    assetCode: { type: String, required: true },
   },
 
   computed: {
@@ -86,10 +91,15 @@ export default {
       businessStatsQuoteAsset: vuexTypes.businessStatsQuoteAsset,
       isAccountCorporate: vuexTypes.isAccountCorporate,
       accountBalanceByCode: vuexTypes.accountBalanceByCode,
+      assetByCode: vuexTypes.assetByCode,
     }),
 
     balance () {
-      return this.accountBalanceByCode(this.asset.code)
+      return this.accountBalanceByCode(this.assetCode)
+    },
+
+    asset () {
+      return this.assetByCode(this.assetCode)
     },
   },
 }